requirements.txt
python-docx
requests
pydub
librosa
soundfile
numpy
pandas
